Place CppCodegen in a separate assembly (#3468)
This is useful for general code hygiene (maybe we should do the same
with RyuJIT and UTC specific stuff), but what forced my hand to do this
now is that we would like to use `ILImporter` for things other than
importing for CppCodegen.

`ILImporter` is a partial class and its reusability depends on users
providing a different part file for the reusable portion. This makes the
class essentially incompatible with any other IL importer implementation
in the same assembly.

I looked into making `ILImporter` reusable, while maintaining the same
perf characteristics (no virtual calls on importing each instruction),
but I couldn't find a scheme where it still made sense.

Maybe we'll end up making the class reusable at some point and absorb
the perf hit, but this pull request has a leg to stand on on it's own.