Make S.L.Expressions.Tests run without asserts. (#3480)
A number of tests was trying to instantiate
a generic type over a generic type definition (List<>).
The typeloader was expressing its understandable
displeasure with an Assert, which stops the text
execution cold (I can only assume Helix is running
DHandler or something similar.)

Nevertheless, on the full framework, this nonsensical
request is allowed - presumably because of another desktop quirk:
the Type representation for List<> (the type definition) is
identical to that of List<T> (List<> instantiated using its own
generic parameter as the type argument.)

.NET Native chose to evolve itself past
this way back in the days of .NETCore 1.0
(keeping the old behavior would have caused
many problems using the .NET Core Reflection
subset api) and we've yet to find a real world
reason to reintroduce it.

I'd rather not even reintroduce this little
bit but allowing the concept merge in the input
direction is less damaging than making them
completely indistinguishable. Furthermore, the
test has hundreds of instances of
this and it'd be hard to explain in a PR why it's
better to fix a 100 lines of test vs a few
lines of product that improves compatibility.