Three Interop bug fixes
This change contains the following bug fixes:
* Call GC.KeepAlive for delegate marshalling. Sergiy encountered a
scenario where the delegate got collected when we tried to do a reverse
P/Invoke. This change adds a GC.KeepAlive at the end of the marshalling
stub to ensure the delegate is not collected during the duration of the
call.
* For ForwardDelegateCreation Stub set the correct calling convention.
Also set the IsPInvoke property of the ForwardDelegateCreation stub to
be true so that the JIT flag CORJIT_FLAG_IL_STUB is set which is
necessary for RyuJit to inline the call.
* Handle the scenarios when MarshalAs can be null.