Switch over to the .NETCore Attribute.Equals() behavior.
This ports https://github.com/dotnet/coreclr/pull/6240 over to Project N.

The TL;DR version of the change is "private fields in base classes are no
longer disregarded by Equals()/GetHashCode()".

In fact, changing ExposeToDerivedClasses() to return true always is sufficient
to effect the desired change. The rest of the delta is just the result of
constant-folding and dead-coding from that point.

[tfs-changeset: 1657635]