Optimize Buffer.BlockCopy for byte arrays
This is similar to the optimization that is done in CoreCLR
(https://github.com/dotnet/coreclr/blob/master/src/vm/comutilnative.cpp#L1379)

Locally I see 30-40% perf improvement on micro-benchmarks with this
changes.

The code in Buffer.BlockCopy can be optimized further (and simplified)
once we have EETypePtr.EETypePtrOf as an inlinable intirisic.