Give RuntimeHelpers.InitializeArray a body
The fact that the method doesn't have a body is a great PITA. E.g. for a long time, creating a delegate to this method would crash the compiler - we now have a transform that generates a throwing body for the method (the transform is intended to sanitize user code though), but that throws a TypeLoadException (for compatibility with the CLR). We should really just have a body for this so that I don't need to special case it in places anymore.

The body will be used only if someone tries to use the method outside of the magical IL sequence (so a reflection call, a delegate, or a handwritten call).

[tfs-changeset: 1658128]