Fix Etw test in System.Collections.Concurrent.tests on ILC (#3621)
Hopefully, this will fix the other Etw-related test failures
I've seen among various projects. ETW propagation to EventListeners
was dying internally with a NullReferenceException on
EventMetadata.Parameters.

EventMetadata.Parameters and EventMetadata.ParameterTypes
represent the same data with EventMetadata.Parameters always
being null on the toolchain-based path and a comment attached to
it saying "Would you hurry up and die already?"

The EventSource rules of the code seems to be to check
before using Parameters and fallback to using ParameterTypes
if necessary but it was forgetting that in a few places.