Address issue 432987 - I looked at a handful of Watson reports, and apparently in the failing cases, the IAT table has been corrupted.
It seems better to fail fast in these cases rather than to AV, or, even worse, crash later in even more confusing ways.

As there are already a bunch of ASSERTs in this code to sanity-check the IAT, it seems safest to turn the ASSERTs into conditional fail fast calls that are active in release build as well.

[tfs-changeset: 1658735]