Fix type loader regression
Checking in on behalf of PeterSol to fix the rolling build:

As Ati noticed, my fix for issue 433565 causes a regression because if the actual type is not a DefType, the cast to DefType obviously fails.

I believe that in this case the type can also not be a nullable type, so one possible way to fix this is to test for the case where the type is not a DefType, and treat it the same way as FixupSignatureKind.TypeHandle if it is not.

[tfs-changeset: 1659248]