Fix 28 System.Runtime.Extensions test failures.
  SetEnvironmentVariable(string, string)
  GetEnvironmentVariables()

  GetEnvironmentVariable(string, target)
  SetEnvironmentVariable(string, string, target)
  GetEnvironmentVariables(target)

   Where target = Target.Process.

   Target.Machine/User will come up next...

Code is mostly cribbed from CoreCLR with some
minor call tree reorg with an eye towards
optimizing out the copies upon copies in
the GetEnvironmentStrings() path.

Code is written with an eye towards
moving to shared partition (though I don't
have time budgeted for that myself.)

Am not doing the Unix versions on CoreRT.
Focus is on getting tests on Project N up.