Implement Assembly.InternalGetSatelliteAssembly to enable ResourceManager.GetString to correctly pull resources from satellite assemblies.
Add Assembly.Load overload that does not throw FileNotFoundExceptions, and make ResourceManager call into that overload.

Add recursion check in SR.cs so that we do not end up doing infinite recursion trying to look up a framework resource, which may happen when the CurrentUICulture is set to a culture that we have not yet localized in S.P.CoreLib.

[tfs-changeset: 1660384]