Remove invalid asserts in the Unix WaitSubsystem WaitObject (#3782)
Fixes #3616.

The crash in #3616 was from a race condition in a small section of code between setting a `WaitHandle` and unregistering a timed out `Wait` call on that `WaitHandle` on a different thread. This section of code is not locked by the `WaitMonitor`. This assert failure happened when a `WaitHandle` was signaled twice while a `Wait` was in the process of timing out between the locked sections.

However, as per @kouvel, the asserts are invalid anyway since when a thread is waiting on multiple `WaitableObject`s (like in `WaitAll`), the thread's wait info will still be in the waiters list of all of the `WaitableObject`s until they all complete. So, we just remove the asserts that were firing since they will break anyway when more multithreaded tests are added.