Enum.GetValues() throws on blocked type.
Fix https://github.com/dotnet/corert/issues/3797

This prompted a rethink of EnumInfo. This was originally
done at a time when we...

- Had only one metadata format to support.
- Far less support (and desire) for caching
  Reflection data.

Today,

- We have three metadata formats to support:

  * NativeFormat
  * EcmaFormat
  * BaloneyFormat (the phony metadata fantasized 
    by blocked types.)

- We've restored strong caching for various
  Reflection data and have the infrastructure for
  it.

Given that, it makes sense now to reimplement
EnumInfo on top of the Type abstraction so as
not to require all this inside knowledge. We're
going to cache the stuff anyway.