Fix https://github.com/dotnet/corert/issues/3800
This implements the "raw" default value apis properly
so that they bypass enum instantation rather than
unwrapping after the fact (which doesn't go back in
time and undo the exception that gets thrown
if you use GetRawConstantValue() to inspect
an open enum.) Also, "raw" is supposed to change
the way we inspect the custom attributes though
we can get away with some simplification here
given our non-support of reflection-only loads.

This exercise also revealed that the Ecma/Native
refactoring for default values was less than
ideal, so this fixes all that up. It also moves
the native half of default value parsing out
of Reflection.Execution and into Reflection.Core
(it was originally there to honor Reflection.Core
coding rules that are no longer applicable.)

Finally, fixed up the "raw" handling of the
CustomConstantAttribute class to match CoreCLR's.
It does create an observable difference even though
we don't support reflection-only loads...