Implement Activator.CreateInstance<T> intrinsic (#3821)
We previously didn't do an intrinsic expanstion for this and had to fall
back to slow and unreliable type loader. This implements a guaranteed
expansion of `CreateInstance<T>` using the infrastructure built for
`EETypePtrOf<T>`.

* Add `RuntimeHelpers.IsReference<T>` intrinsic that constantly expands
to 0 or 1 at compile time (this gets rid of the `default(T) == null`
workaround that requires special handling for `Nullable<T>`).
* Add `DefaultConstructorOf<T>` that expands to the default constructor
of `T` or a special marker method.
* Rewrite `CreateInstanceIntrinsic` using the new intrinsics.

Fixes #368.