Fix https://github.com/dotnet/corert/issues/3810
Seeing "$BlockedFromReflection_3_19f5285d" rather than
"MissingMetadataEsxception" in xunit test logs is getting
to be a bore. The reducer already enables browsing for
all Exception types to avoid this kind of eyesore -
internally blocked types that nevertheless love to
get outside the framework at every opportunity
need a little help for this.

[tfs-changeset: 1660982]