[WIP] Unify API between StackFrameIterator and various special addresses
In .NET Native, we're hitting an issue where DIA gets confused by
public code labels pointing into the middle of methods as it
messes up stack frame layout. I fixed this before for
the UniversalTransitionThunk and for CallDescrWorker. This change
extends the same treatment to ExceptionHandlers and furthermore
introduced a common cross-architecture macro to unify this idiom.