UnmanagedFunctionPointer Attribute
We were handling  only CallingConvension from UnmanagedFuncPointerAttrbiutes , but the networking team hit an issue where they were relaying on SetLastError on by default. This change implement SetLastError , but it's not on by default as in CoreCLR/Desktop. Now we read and store all the attributes from UnmanagedFunctionPointer but only CallingConvesion and SetLastError are handled.

Change also include a fix for DllImport SetLastError whihc was not clearing the error before setting it.

[tfs-changeset: 1661289]