Allow an external optimizer to specify VTable layout (#3854)
Adds an extensibility point to the `CompilationBuilder` that allows
specifying how vtables should be built.

We currently have two strategies: eagerly built VTables (where we just
put all the virtual methods that the type specifies in the metadata into
the VTable), and lazily built vtables (where we track virtual method
usage during the compilation and only generate vtable slots for methods
that were actually used).

The advantage of the former is that we know the slot assignment very
early on and can inline it into the codegen. The advantage of the latter
is natural reduction of the amount of code we compile.

I'm adding an option to specify optional optimization data that can be
used instead of the lazily built strategy and making it possible for the
IL scanner to generate this data. This way we can have an option where
we both have the natural reduction in the amount of generated code, and
also the inlining of slot numbers in codegen (that part depends on #3773
to light up.