Cleanups after enabling virtual method inlining
Pull request #3773 enabled some cleanups I didn't want to do there
because it's a rather big diff with a possible (expected) test
regression:

* ReadyToRunHelperId.VirtualCall and
ReadyToRunHelperId.ResolveVirtualFunction should no longer show up in
connection with generic lookups. We only used them for interface calls
and for those, RyuJIT now knows how to handle with dispatch cells
directly. They're still used in non-generic codepaths where they handle
normal interface calls with unknown slot numbers.
* Standalone LDVIRTFTN for interface methods on generic interfaces used
a hack where we placed a R2R helper into the dictionary and returned
that. Given this is only reachable outside of delegate creation, I'm
deleting the hack. If we need it, we should implement it the right way,
because this returns function pointers to things that are not user code.
It works most of the time, but not always. I'm expecting some rolling
build fallout from this.
* I'm deleting one of the hacky generic dictionary entries. We should
rename the second one, but that's best done from TFS.