Support TypeHandleFixupAttribute and MethodAddrFixupAttribute (#3858)
The code adds TypeHandleFixup/MethodAddrFixup and abstraction for fixups. It is mostly straight-forward but testing is rather annoying:

The data blobs for fixups are pointer-size and therefore dependent on CPU architecture. I haven't found anything in ILAsm that is similar to ifdef so I have to conditionally include 32-bit or 64-bit copies. I suppose we'll need to deal with the fun that is indian-ness at some point.
Updating IL is quite annoying as I don't want to write test code in IL. I end up splitting the IL in two sections, and I compile the corresponding C# code into exe, ildasm it, and incrementally copy/paste the code that I want into the IL.
I need to use internal types such as FixupRuntimeTypeHandle / NativeCallable, so those need to be manually fixed up from my internal copy to the right reference System.Private.CoreLib