Stop passing around redundant parameters
Project X introduced dictionary cells whose targets depend on the
generic dictionary they're contained in. This required adding a separate
`GenericDictionary` parameter to `GetTarget`. CoreRT wasn't fixed up to
provide it and we ended up with redundant parameters and a null
dictionary there.

Instead of providing the dictionary (which would require multiple hash
table lookups to obtain per each generic dictionary cell during the
dependency analysis phase), I'm fixing this up to provide enough
breadcrums so that the single dictionary cell that really cares can get
to the dictionary and compute it when needed.