Generate more interface dispatch cells
This is still not "per callsite", but at least per method/generic
dictionary.

We could make it per callsite with either some RyuJIT support, or by
scanning relocs generated by RyuJIT and disambiguating the references to
dispatch cells (another option would be to generate a new dispatch cell
per each `getCallInfo` call, but that would produce a bunch of garbage
that will never be marked in the graph).