Delete ToolsVersion attribute from msbuild files
The latest project templates do not have it, and it is only good for generating warnings like the following in the detailed build log:

Project file contains ToolsVersion="4.0". This toolset may be unknown or missing, in which case you may be able to resolve this by installing the appropriate version of MSBuild, or the build may have been forced to a particular ToolsVersion for policy reasons. Treating the project as if it had ToolsVersion="14.0". For more information, please see http://go.microsoft.com/fwlink/?LinkId=293424.

[tfs-changeset: 1662024]