Further improve perf of Char.IsWhiteSpace (dotnet/coreclr#12353)
IsWhiteSpaceLatin1 is not being inlined.  By changing a range check done with two comparison operations to instead be done with a subtraction and a single comparison, the code is shortened to not only be less expensive but also then get inlined into IsWhiteSpace and then further into String.IsNullOrWhiteSpace.  The net result is a measurable throughput improvement for IsNullOrWhiteSpace.

Signed-off-by: dotnet-bot <dotnet-bot@microsoft.com>