Fix to GVM resolution bug.
VB can generate calls to GVMs using the signature of a derived type, instead of the type where the GVM is declared (like in C#). Our toolchain/GVM resolution logic could not handle this.

The fix is to detect this and return the declaring method from the JIT interface, and encode its RMH in the GVM lookup (this will have the right declaring type, which will match the way we write the entries in the GVM tables)

[tfs-changeset: 1662438]