Fix for bug 453367: unhandled MissingTemplateExceptions thrown from the TypeLoader on ARM.
Creating dynamic types from non-USG templates, where the USG version of the template did not exist was throwing the exception. The TypeLoader needed to compute the HFA flag, and to do so, needed to compute the field layout from the USG template (which didn't exist).
The fix is to compute the HFA flag from the non-USG template directly without having to compute the field layout (information stored in rare flags on the template EEType). Field layout is still fully computed when creating dynamic types from USG templates.

[tfs-changeset: 1662503]