Fix failing System.Runtime.Tests on ILC (Sort_Array_Array_NonGeneric)
Fixes https://github.com/dotnet/corefx/issues/17450

This test got reenabled recently though the underlying
bug wasn't fixed. The test was still hitting an assert in CoreLib.

Fixing this now.

It's perfectly legal for the keys and values arrays passed to Array.Sort()
to be of different sizes. All that's required is that offset/range you're
sorting fits within both.

So the "clever" shortcut this assertion rack uses to avoid writing
a proper range check against the "values" array is not valid.