Clean up serialization from Encoding
We have moved more Encoding code to the shared folder and it had some serialization code which is not needed. The change here is cleaning up these changes and update the field names back.
Note, I kept the types that implement ISrializable interface to match the desktop contracts to not break anything