Overhaul exception throwing from the type system
Throwing an exception from the type system is a rather complicated thing because the type system is used in so many places. The basic requirements are:

* The runtime type system needs to throw the runtime exceptions (System.TypeLoadException, System.BadImageFormatException, etc.).
* The compile time type system needs to throw exceptions in a way that we can "serialize" the thrown exception into a chunk of code that will throw the desired exception at runtime. This is to emulate the CLR behavior you would get if you e.g. mess up your StructLayout, or feed the compiler garbage in general.
* The exception message thrown or serialized from the compiler can have different languages (the localization of the compiler doesn't have to match that of the running compiled executable).

With this change, all exception throwing happens through a ThrowHelper. We have two ThrowHelpers - one used by "compile time type systems" and another one used by the runtime type system. They share the logic to format the type names and exception message strings.

[tfs-changeset: 1665419]