Fix for 462734
DelegateMethodInfoRetriever.GetDelegateMethodInfo was recently changed to throw MissingRuntimeArtifactException instead of returning null.

This caused one of the DG test to fail, and uncovered an actual real bug. The test was performing delegate.GetMethodInfo() calls, and asserting that 2 delegates constructed in different ways were returning the same MethodInfo. The test was passing for the wrong reason: because null == null is true :). The test needed some rd.xml entries to make the delegate's target method reflectable in order to get a valid MethodInfo.

In the USG case, delegate targets could be wrapped inside call converter thunks, which wasn't detected/handled correctly by the logic in TryGetMethodForOriginalLdFtnResult.

[tfs-changeset: 1665453]