Fix memory consistency issue in ObjectHeader.GetHashCode
With selfhosting roslyn we are recently seeing a plethora of strange errors in appcompat and other test infrastructures. The parallel nature of roslyn helped us to unearthed this bug where in ObjectHeader.GetHashCode we dereference a pointer and check the value of it. Compiler  generates code to dereference the pointer twice and in multi-threaded scenarios some other thread rewrites the memory pointed by the header. The fix is to do a volatile.read. For arm where volatile.read is quite expensive as it involves a barrier, instead we do a function call with noinlining to ensure the pointer is not dereferenced again.

[tfs-changeset: 1667190]