Add tests for dependency graph
This adds a unit test for `ILCompiler.Compiler` so that we can no longer
use this as an excuse not to write unit tests for this assembly.

The first test is a generalized framework to write tests for dependency
graph. The test runs ILScanner on a method from the test assembly, and
validates various invariants that the method declares. This will let us
e.g. write targeted unit tests that check that the various size on disk
features within the compiler do the right thing (e.g. make sure we don't
unnecessarily expand variant interfaces, GVMs, etc.).