Fix MethodEntry dictionary cells to point to the correct unboxing stub pointer.
Also fix the native layout to capture this information (modulo the USG case which needs the non-unboxing stub to avoid unnecessary double thunking. See binder code in GenericDictionary.cpp. In the case of USG, we only store a flag indicting unboxing is needed, but we store the pointer to the non-unboxing stub).

Adding more tests to ProjectXDynamicGenerics

[tfs-changeset: 1669569]