Improve diagnosability of Expression.Lambda(Expression.Constant(123)).Compile().GetMethodInfo()
Entity framework would like to inspect the MethodInfo obtained from a delegate that points to the LINQ expression interpreter. This will currently throw a very unhelpful exception that makes it look like we're missing metadata - when in fact, we're missing a feature.

EF will ship with a workaround in their 2.0.1 release, but if anyone tries with 2.0.0 (or hits this in any other way), let's give them something more diagnosable.

[tfs-changeset: 1670903]